SAK-49345 Switch from Base62 to a safer Base64
