+++ basiclti/basiclti-tool/src/java/org/sakaiproject/blti/tool/LTIAdminTool.java
@@ -57,7 +57,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import org.json.simple.JSONObject;
 import org.json.simple.JSONArray;
 
-import org.tsugi.util.Base62;
+import org.tsugi.util.Base64DoubleUrlEncodeSafe;
 import static org.tsugi.basiclti.BasicLTIUtil.getObject;
 import static org.tsugi.basiclti.BasicLTIUtil.getString;
 
@@ -1673,7 +1673,7 @@ public class LTIAdminTool extends VelocityPortletPaneledAction {
 		}
 
 		// Sanity check our (within Sakai) returnUrl
-		String returnUrl = Base62.decode(data.getParameters().getString("returnUrl"));
+		String returnUrl = Base64DoubleUrlEncodeSafe.decode(data.getParameters().getString("returnUrl"));
 		if (returnUrl == null) {
 			addAlert(state, rb.getString("error.contentitem.missing.returnurl"));
 			switchPanel(state, errorPanel);
@@ -2528,7 +2528,7 @@ public class LTIAdminTool extends VelocityPortletPaneledAction {
 				+ "?eventSubmit_doSingleContentItemResponse=Save"
 				+ "&" + FLOW_PARAMETER + "=" + flow
 				+ "&" + RequestFilter.ATTR_SESSION + "=" + URLEncoder.encode(sessionid + "." + suffix)
-				+ "&returnUrl=" + Base62.encode(returnUrl)
+				+ "&returnUrl=" + Base64DoubleUrlEncodeSafe.encode(returnUrl)
 				+ "&panel=PostContentItem"
 				+ "&tool_id=" + tool.get(LTIService.LTI_ID);
 
@@ -2558,7 +2558,7 @@ public class LTIAdminTool extends VelocityPortletPaneledAction {
 
 		// Run the contentreturn through the forward servlet
 		contentReturn = serverConfigurationService.getServerUrl() + "/imsoidc/lti13/resigncontentitem?forward=" +
-			Base62.encode(contentReturn) + "&tool_id=" + tool.get(LTIService.LTI_ID);
+			Base64DoubleUrlEncodeSafe.encode(contentReturn) + "&tool_id=" + tool.get(LTIService.LTI_ID);
 
 		contentLaunch = ContentItem.buildLaunch(contentLaunch, contentReturn, contentData);
 
@@ -2796,7 +2796,7 @@ public class LTIAdminTool extends VelocityPortletPaneledAction {
 					+ "/sakai.lti.admin.helper.helper"
 					+ "?panel=ContentConfig"
 					+ "&" + FLOW_PARAMETER + "=" + flow
-					+ "&returnUrl=" + Base62.encode(returnUrl)
+					+ "&returnUrl=" + Base64DoubleUrlEncodeSafe.encode(returnUrl)
 					+ "&tool_id=" + tool.get(LTIService.LTI_ID)
 					+ "&" + RequestFilter.ATTR_SESSION + "=" + URLEncoder.encode(sessionid + "." + suffix);
 			context.put("forwardUrl", configUrl);
@@ -2840,7 +2840,7 @@ public class LTIAdminTool extends VelocityPortletPaneledAction {
 
 		// Run the contentreturn through the forward servlet
 		contentReturn = serverConfigurationService.getServerUrl() + "/imsoidc/lti13/resigncontentitem?forward=" +
-			Base62.encode(contentReturn) + "&tool_id=" + tool.get(LTIService.LTI_ID);
+			Base64DoubleUrlEncodeSafe.encode(contentReturn) + "&tool_id=" + tool.get(LTIService.LTI_ID);
 
 		// This will forward to AccessServlet / BasicLTISecurityServiceImpl with a tool: url
 		// AccessServlet will detect if this is a CI or DL and handle it accordingly using
